New functions. Cleaner patient handling.

+++ main.py
@@ -9,7 +9,7 @@ from pathlib import Path
 import sys
 import matplotlib.pyplot as plt
 import numpy as np
-
+import cv2
 from PySide2.QtWidgets import QApplication, QMainWindow, QGraphicsView, QLabel, QPushButton, QTextBrowser, QAction, QFileDialog, QDialog, QDialog
 from PySide2.QtCore import QFile, QObject, SIGNAL, QDir 
 from PySide2.QtUiTools import QUiLoader 
@@ -17,7 +17,6 @@ from PySide2.QtGui import QImage, QPixmap
 from segment import Image_2_seg
 from manualseg import manualSeg
 
-import cv2
 
 class Window(QMainWindow):
     def __init__(self):
@@ -31,6 +30,14 @@ class Window(QMainWindow):
         self.inputExists = False
         self.defaultDirectoryExists = False
 
+    def load_ui(self):
+        loader = QUiLoader()        
+        path = os.fspath(Path(__file__).resolve().parent / "form.ui")
+        ui_file = QFile(path)
+        ui_file.open(QFile.ReadOnly)
+        self.ui_window = loader.load(ui_file, self)
+        ui_file.close()
+    
     def make_connect(self):
         QObject.connect(self.ui_window.segButton, SIGNAL ('clicked()'), self.segment)
         QObject.connect(self.ui_window.tempButton, SIGNAL ('clicked()'), self.temp_extract)
@@ -42,19 +49,15 @@ class Window(QMainWindow):
         QObject.connect(self.ui_window.previousImageButton , SIGNAL ('clicked()'), self.previousImage)
         QObject.connect(self.ui_window.saveButton , SIGNAL ('clicked()'), self.saveImage)
 
-    def load_ui(self):
-        loader = QUiLoader()        
-        path = os.fspath(Path(__file__).resolve().parent / "form.ui")
-        ui_file = QFile(path)
-        ui_file.open(QFile.ReadOnly)
-        self.ui_window = loader.load(ui_file, self)
-        ui_file.close()
-
-    def messagePrint(self):
+    def messagePrint(self, message):
         #INPUT: string to print
         #OUTPUT: none
-        #ACTION. generate out.html file and refresh it in Messages QTextArea
-        pass
+        #ACTION: generate out.html file and refresh it in Messages QTextArea
+        out_file = open("out.html" , "w")
+        out_file.write(message)
+        out_file.close()
+        self.ui_window.textBrowser.setSource('out.html')
+        self.ui_window.textBrowser.reload()
 
     def findImages(self):
         self.fileList = []
@@ -83,24 +86,19 @@ class Window(QMainWindow):
         self.i2s = Image_2_seg()
         self.i2s.setPath(self.opdir)
         self.i2s.extract()
-        print(self.i2s.Y_pred)
         threshold =  0.5
         img = plt.imread(self.opdir)/255
-        Y = self.i2s.Y_pred / self.i2s.Y_pred.max()
+        Y = self.i2s.Y_pred
+        Y = Y / Y.max()
         Y = np.where( Y >= threshold  , 1 , 0)
-        Y = cv2.resize(Y[0] , (self.inputImage.shape[1] , self.inputImage.shape[0] ) , interpolation = cv2.INTER_NEAREST)
-        Y = cv2.cvtColor(Y , cv2.COLOR_BGR2RGB)
-        cv2.imwrite("output.jpg" , Y)
+        Y = cv2.resize(Y[0], (img.shape[1],img.shape[0]), interpolation = cv2.INTER_NEAREST) # Resize the prediction to have the same dimensions as the input 
+        #Y_im = cv2.cvtColor(Y, cv2.COLOR_BGR2RGB)
+        #cv2.imwrite("output.jpg", Y_im) 
+        plt.imsave("output.jpg" , Y*0.5+img[:,:,0] , cmap='gray')
         self.ui_window.outputImg.setPixmap("output.jpg")
 
-
     def segment(self):
-        out_file = open("out.html" , "w")
         if self.inputExists:
-            out_file.write("Imagen segmentada exitosamente")
-            out_file.close()
-            self.ui_window.textBrowser.setSource('out.html')
-            self.ui_window.textBrowser.reload()
             self.feet_segment()
         else:
             out_file.write("No se ha seleccionado imagen de entrada")
@@ -115,19 +113,13 @@ class Window(QMainWindow):
         return
 
     def temp_extract(self):
-        out_file = open("out.html" , "w")
         if self.inputExists:
-            out_file.write("Se extrajo temperatura exitosamente")
-            out_file.close()
-            self.ui_window.textBrowser.setSource('out.html')
-            self.ui_window.textBrowser.reload()
+            self.messagePrint("Se extrajo la temperatura exitosamente")
         else:
-            out_file.write("No se ha seleccionado imagen de entrada")
-            out_file.close()
-            self.ui_window.textBrowser.setSource('out.html')
-            self.ui_window.textBrowser.reload()
+            self.messagePrint("No se ha seleccionado un imagen de entrada")
     
     def figlabels(self):
+        #  Get info from directory path name and obtain time indexes based on name
         pass
 
     def openImage(self):
@@ -160,19 +152,22 @@ class Window(QMainWindow):
             self.findImages()
 
     def makeTimePlot(self):
-        x=np.array([0,1,5,10,15,20])
-        y=np.array([35.5, 35.7 , 36 , 37.2 , 37.3, 37.5])
-        fig=plt.figure(figsize=(9.6,4))
-        plt.plot(x,y,label='Paciente 1')
-        plt.legend()
-        plt.grid()
-        plt.xlabel("Tiempo [minutos]")
-        plt.ylabel("Temperatura [°C]")
-        plt.title("Time plot")
-        #plt.show()
-        plt.savefig('fresh.png')
-        self.ui_window.timePlot.setPixmap('fresh.png')
-
+        if self.inputExists:
+            x=np.array([0,1,5,10,15,20])
+            y=np.array([35.5, 35.7 , 36 , 37.2 , 37.3, 37.5])
+            fig=plt.figure(figsize=(9.6,4))
+            plt.plot(x,y,label='Paciente 1')
+            plt.legend()
+            plt.grid()
+            plt.xlabel("Tiempo [minutos]")
+            plt.ylabel("Temperatura [°C]")
+            plt.title("Time plot")
+            #plt.show()
+            plt.savefig('fresh.png')
+            self.ui_window.timePlot.setPixmap('fresh.png')
+            self.messagePrint("Se ha actualizado el TimePlot")
+        else:
+            self.messagePrint("No se puede actualizar el TimePlot. No se ha seleccionado una imagen de entrada")
 
 if __name__ == "__main__":
     app = QApplication(sys.argv)
